Extract llg_test_utils crate and move tests to parser
Create llg_test_utils workspace-only crate (publish=false) by merging
sample_parser/src/lib.rs and tests/common_lark_utils/mod.rs into shared
test infrastructure.

Move all 12 integration test files from sample_parser/tests/ to
parser/tests/ (standard Rust integration test location). Update imports
from sample_parser::*/common_lark_utils::* to llg_test_utils::*.

Simplify sample_parser to purely demo binaries + data files by removing
lib.rs, tests/, and dev-dependencies.
