+++ Cargo.toml
@@ -4,6 +4,7 @@ members = [
     "parser",
     "sample_parser",
     "json_stats",
+    "llg_test_utils",
     "toktrie",
     "toktrie_hf_tokenizers",
     "toktrie_hf_downloader",
@@ -41,6 +42,7 @@ opt-level = 3
 [workspace.dependencies]
 toktrie = { path = "toktrie" }
 llguidance = { path = "parser" }
+llg_test_utils = { path = "llg_test_utils" }
 toktrie_hf_tokenizers = { path = "toktrie_hf_tokenizers" }
 toktrie_hf_downloader = { path = "toktrie_hf_downloader" }
 toktrie_tiktoken = { path = "toktrie_tiktoken" }
