+++ parser/tests/test_bias_cache.rs
@@ -13,7 +13,7 @@ use llguidance::{
     toktrie::{InferenceCapabilities, SimpleVob},
     Matcher, ParserFactory,
 };
-use sample_parser::get_tok_env;
+use llg_test_utils::get_tok_env;
 
 lazy_static! {
     static ref PARSER_FACTORY: ParserFactory = {
