+++ parser/tests/test_json_arrays.rs
@@ -2,8 +2,7 @@ use lazy_static::lazy_static;
 use rstest::*;
 use serde_json::{json, Value};
 
-mod common_lark_utils;
-use common_lark_utils::{json_err_test, json_schema_check};
+use llg_test_utils::{json_err_test, json_schema_check};
 
 lazy_static! {
     static ref INTEGER_ARRAY: Value = json!({"type":"array", "items": {"type":"integer"}});
